requirements.txt
# apache-airflow-providers-snowflake==3.2.0
# snowflake-connector-python==2.7.12
apache-airflow==2.3.4
snowflake-sqlalchemy==1.4.1
pandas==1.4.4
